Improved large record handling in RAW format
The original version was saving all string and blob arguments with 16-bit length.
That is usually enough for the strings but is not for large blobs. Also the formatter
was not handling truncation properly (effectively casting unsigned long to uint16).

This commit fixes those issues and introduces support for legacy RAW versions in
the raw-parser and hogl-cook tool.