Fix stack buffer overflow of variable-length array
Steps to reproduce on Ubuntu 14.04:

Install a toolchain with a recent version of the address sanitizer (gcc 4.8 is too old):

$ sudo apt-add-repository "deb http://llvm.org/apt/trusty/ llvm-toolchain-trusty-3.8 main"
$ sudo apt-get update
$ sudo apt-get install clang-3.8

$ make -j CC=clang-3.8 CXX=clang++-3.8 CXXFLAGS="-g -fsanitize=address" LDFLAGS="-g -fsanitize=address"

Disable less evil errors and tell ASan where to find llvm-symbolizer
so that you get a nice, readable stack trace:

$ ASAN_OPTIONS="detect_odr_violation=0 alloc_dealloc_mismatch=0" PATH=$PATH:/usr/lib/llvm-3.8/bin make check

View the .log files of the failed tests. Or run the failing test by hand
(with the same environment variables set) to see a colorful error message on stdout.