Make tests and tools optional
  --without-tools         Do not build tools
  --without-tests         Do not build tests

The build currently doesn't offer a way to run a cross-compiled test suite,
so add a way to disable it. This might also be useful if a cross-compiled version
of boost is unavailable.

Also add a way to disable building the tools, which may not be useful in the
cross-compiled environment or not be buildable if a cross-compiled binutils-dev
is unavailable.

To cross-compile for ARM, ensure <PREFIX>-gcc is in your system path and then run:

   ./configure --without-tools --without-tests --host=<PREFIX>

where <PREFIX> might be 'arm-linux-gnueabihf'