Change hogl::tls to remember the previous and the current TLS ring buffers.
Previously hogl::tls used thread local variable to remember the ring buffer
created and used. This is problematic when multiple hogl::tls instances are
created in the same thread, as subsequent hogl::tls instance overwrites the
same thread local variable with its own ring buffer on construction and
restore it to the default ring buffer on destruction. The result is that
ring buffer created by latter hogl::tls instances are not dereference counted.

With this change, dedicated member variables _previous_ring and _current_ring
are used to remember the previous and the current TSL ring buffer, the TLS
varible _ring is not relied on to be the current TLS ring buffer.