Support for tee output and minor cleanup of the output interfaces
It was always possible to write a custom output handler that implements
the tee functionaility but it wasn't as straightforward as it should be
and required some redundant buffering and things.

This commit introduces new hogl::output_tee and hogl::ostrbuf_tee that
allow for clean output t-ing without extra overhead and buffering.