Fix longstanding inconsitency in record_tailroom definition
ringbuf implementation and ringbuf::record_tailroom() method always assumed that the tailroom
includes argument storage, basically the full record minus the record::header_size().
But the tailroom requested via ringbuf::options was simply added to the full record which resulted
in overallocation, especially when the power-of-two rounding of the record size is taken into account.

This commit fixes the interpretation of the tailroom in ringbuf::options to match the rest of the
ringbuf logic.