Allow spaces in format validator and make hogl-cook more configurable
Spaces should be allowed at this layer since nothing else in processing logic
is affected. It's not a good idea to use them for ring & section names but it
is up to the users.

This commit also makes hogl-cook more configurable with --min-valid-recs option.
The option sets min number of valid records to expect in the .raw file to avoid
retrying validation from the beginning.