Fix incorrect type used in bitmap shift
When the word size is 64 bits, the value being shifted to test a bit
needs to be a 64-bit value. Use a static_cast to force the value to the
word type so that there is no undefined behavior (a bit shift greater
than or equal to the length in bits of the value being shifted is
undefined behavior).